# backend/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API for ingesting, storing, and querying log data with Grok pattern support",
        "title": "LogSonic API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
...
        "/ingest/jobs/{id}": {
            "delete": {
                "description": "Cancel a running path-based ingest job. A no-op (200, the job's own terminal state) if it already finished.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "ingest"
                ],
                "summary": "Cancel a path-ingest job",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Job ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/types.IngestJobActionResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/types.ErrorResponse"
                        }
                    }
                }
            }
        },
...
        "/sources": {
            "get": {
                "description": "Every source in the catalog (\u003cstorage\u003e/sources.json): origin, pattern, rows, bytes, timestamp span, the day-indices holding it, and its import history. Maintained on the ingest write path; see POST /sources/rebuild to reconcile with the indices.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "sources"
                ],
                "summary": "List sources",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/types.SourcesResponse"
                        }
                    },
                    "503": {
                        "description": "Service Unavailable",
                        "schema": {
                            "$ref": "#/definitions/types.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/sources/rebuild": {
            "post": {
                "description": "Recomputes rows, days and timestamp bounds for every source from the day-indices (one size-0 facet per day on current indices; a stored-field read on indices created before v1.8's keyword _src mapping). Origin, pattern and import history are kept; sources with no rows left are dropped. Idempotent.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "sources"
                ],
                "summary": "Rebuild the sources catalog from the indices",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/types.SourcesResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/types.ErrorResponse"
                        }
                    },
                    "503": {
                        "description": "Service Unavailable",
                        "schema": {
                            "$ref": "#/definitions/types.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/sources/{name}": {
            "get": {
                "description": "One catalog entry by source name (the stored _src value).",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "sources"
                ],
                "summary": "Get one source",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Source name",
                        "name": "name",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/types.SourceEntry"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/types.ErrorResponse"
                        }
                    },
                    "503": {
                        "description": "Service Unavailable",
                        "schema": {
                            "$ref": "#/definitions/types.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "types.ErrorResponse": {
            "description": "Standardized error response structure used across all API endpoints",
            "type": "object",
            "properties": {
                "code": {
                    "description": "Error code for programmatic handling",
                    "type": "string"
                },
                "details": {
                    "description": "Additional error details",
                    "type": "string"
                },
                "error": {
                    "description": "Main error message",
                    "type": "string"
                },
                "status": {
                    "description": "Status will always be \"error\" for error responses",
                    "type": "string"
                }
            }
        },
...
        "types.IngestJobActionResponse": {
            "type": "object",
            "properties": {
                "status": {
                    "description": "\"cancelling\", or the job's terminal state if it had already finished",
                    "type": "string"
                }
            }
        },
...
        "types.SourceEntry": {
            "type": "object",
            "properties": {
                "bytes_raw": {
                    "type": "integer"
                },
                "created_at": {
                    "type": "string"
                },
                "day_rows": {
                    "type": "object",
                    "additionalProperties": {
                        "type": "integer",
                        "format": "int64"
                    }
                },
                "days": {
                    "description": "Days lists every day-index holding rows for this source (sorted);\nDayRows is the per-day count behind it, which is what lets a rebuild\nafter a prune or a row delete touch only the affected days.",
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "first_ts": {
                    "type": "string"
                },
                "imports": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/types.SourceImport"
                    }
                },
                "last_ts": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "origin": {
                    "$ref": "#/definitions/types.SourceOrigin"
                },
                "pattern": {
                    "type": "string"
                },
                "pattern_name": {
                    "type": "string"
                },
                "rows": {
                    "type": "integer"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
        "types.SourceImport": {
            "type": "object",
            "properties": {
                "at": {
                    "type": "string"
                },
                "job_id": {
                    "type": "string"
                },
                "path": {
                    "type": "string"
                },
                "rows": {
                    "type": "integer"
                }
            }
        },
        "types.SourceOrigin": {
            "type": "object",
            "properties": {
                "host": {
                    "type": "string"
                },
                "kind": {
                    "type": "string"
                },
                "path": {
                    "type": "string"
                }
            }
        },
...
        "types.SourcesResponse": {
            "type": "object",
            "properties": {
                "sources": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/types.SourceEntry"
                    }
                }
            }
        },
...
    }
}